+++ ext/geos_c_impl/main.c
@@ -2,7 +2,6 @@
   Main initializer for GEOS wrapper
 */
 
-
 #include "preface.h"
 
 #ifdef RGEO_GEOS_SUPPORTED
@@ -10,17 +9,19 @@
 #include <ruby.h>
 #include <geos_c.h>
 
+#include "errors.h"
+
 #include "factory.h"
 #include "geometry.h"
 #include "point.h"
 #include "line_string.h"
 #include "polygon.h"
 #include "geometry_collection.h"
+#include "analysis.h"
 
 #endif
 
 RGEO_BEGIN_C
-
 
 void Init_geos_c_impl()
 {
@@ -33,6 +34,8 @@ void Init_geos_c_impl()
   rgeo_init_geos_line_string(globals);
   rgeo_init_geos_polygon(globals);
   rgeo_init_geos_geometry_collection(globals);
+  rgeo_init_geos_analysis(globals);
+  rgeo_init_geos_errors();
 #endif
 }