+++ lib/rgeo/impl_helper/basic_geometry_collection_methods.rb
@@ -114,6 +114,12 @@ module RGeo
         @elements.map(&:coordinates)
       end
 
+      def contains?(rhs)
+        return super unless Feature::Point === rhs
+
+        @elements.any? { |line| line.contains?(rhs) }
+      end
+
       private
 
       def add_boundary(hash, point)
@@ -176,6 +182,12 @@ module RGeo
 
       def coordinates
         @elements.map(&:coordinates)
+      end
+
+      def contains?(rhs)
+        return super unless Feature::Point === rhs
+
+        @elements.any? { |poly| poly.contains?(rhs) }
       end
     end
   end