+++ test/common/line_string_tests.rb
@@ -331,6 +331,27 @@ module RGeo
           line = @factory.line_string([point1, point2, point3])
           assert_equal(line.point_on_surface, point2)
         end
+
+        def test_contains_point
+          point1 = @factory.point(0, 0)
+          point2 = @factory.point(1, 2)
+          point3 = @factory.point(3, 9)
+          line_string = @factory.line_string([point1, point2, point3, point1])
+
+          assert_equal(true, line_string.contains?(@factory.point(0.5, 1)))
+          assert_equal(false, line_string.contains?(@factory.point(-1, -1)))
+        end
+
+        def test_linear_ring_ccw
+          point1 = @factory.point(1, 1)
+          point2 = @factory.point(2, 0)
+          point3 = @factory.point(0, 0)
+          cw_ring = @factory.linear_ring([point1, point2, point3, point1])
+          ccw_ring = @factory.linear_ring([point3, point2, point1, point3])
+
+          assert_equal(false, cw_ring.ccw?)
+          assert_equal(true, ccw_ring.ccw?)
+        end
       end
     end
   end