+++ test/test_helper.rb
@@ -1,15 +1,15 @@
 # frozen_string_literal: true
 
 require "minitest/autorun"
-require "rgeo"
+require_relative "../lib/rgeo"
 require "psych"
-require "common/factory_tests"
-require "common/geometry_collection_tests"
-require "common/line_string_tests"
-require "common/multi_line_string_tests"
-require "common/multi_point_tests"
-require "common/multi_polygon_tests"
-require "common/point_tests"
-require "common/polygon_tests"
+require_relative "common/factory_tests"
+require_relative "common/geometry_collection_tests"
+require_relative "common/line_string_tests"
+require_relative "common/multi_line_string_tests"
+require_relative "common/multi_point_tests"
+require_relative "common/multi_polygon_tests"
+require_relative "common/point_tests"
+require_relative "common/polygon_tests"
 
 require "pry-byebug" if ENV["BYEBUG"]